Code/ZhiJinCompiler/3rdparty/AutoBuild/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
bidict==0.22.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dnspython==2.3.0
eventlet==0.33.3
exceptiongroup==1.1.1
executing==1.2.0
Flask==2.3.2
Flask-SocketIO==5.3.4
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
outcome==1.2.0
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
PySocks==1.7.1
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
requests==2.30.0
selenium==4.9.0
simple-websocket==0.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2
